feat: Print human-readable output if stdout is TTY
rust-analyzer requires us to print to stdout in a very particular format
so it can understand what we're saying. Unfortunately, this means that
running `cargo-subspace` on the command line yourself can be quite a
dizzying experience. This commit improves this situation by checking if
we're running in a TTY. If we are, we make the following changes:
- The default log level becomes `INFO` instead of `WARN` (since we don't
  have to worry about filling up the filesystem)
- We log progress events as human-readable strings with `tracing::info!`
- We pretty print the project JSON