bug: Loosen file path requirement for discover command
Previously, we were expecting the path passed to `DiscoverArgument::Path`
to be path to a file that already exists. However, it's possible that
rust-analyzer will invoke us with a path to a not-yet-existing file if
the user has opened a new file in their editor but hasn't yet saved.
This commit removes the requirement that this path point to a file on
disk to improve UX in this situation.